PythonInterface/requirements.txt
pillow
pillow_heif==0.15.0
pyinstaller